+++ mainprogs/tests/t_dwflinop.cc
@@ -32,7 +32,7 @@ void readSzinFerm(multi1d<LatticeFermion>& q, const std::string& file)
   {
     for(int cb=0; cb < 2; ++cb)
     {
-      for(int sitecb=0; sitecb < Layout::vol()/2; ++sitecb)
+      for(size_t sitecb=0; sitecb < Layout::vol()/2; ++sitecb)
       {
 	multi1d<int> coord = crtesn(sitecb, lattsize_cb);
